+++ tests/gradients/test_pulse_gradient.py
@@ -16,10 +16,23 @@ Note that the module is implemented for the new return type system
 so that this test suite just tests for correctly raised errors."""
 import pytest
 import pennylane as qml
+from pennylane.gradients.pulse_gradient import stoch_pulse_grad
 
-
+@pytest.mark.jax
 def test_stoch_pulse_grad_raises():
     """Test that stoch_pulse_grad raises a NotImplementedError."""
     tape = qml.tape.QuantumScript()
     with pytest.raises(NotImplementedError, match="The stochastic pulse parameter-shift"):
         qml.gradients.stoch_pulse_grad(tape)
+
+def test_stoch_pulse_grad_raises_without_jax_installed():
+    """Test that an error is raised if a stoch_pulse_grad is called without jax installed"""
+    try:
+        import jax  # pylint: disable=unused-import
+
+        pytest.skip()
+    except ImportError:
+        tape = qml.tape.QuantumScript([], [])
+        with pytest.raises(ImportError, match="Module jax is required"):
+            stoch_pulse_grad(tape)
+
