+++ tests/returntypes/gradients/test_pulse_gradient_new.py
@@ -24,19 +24,6 @@ import pennylane as qml
 from pennylane.gradients.pulse_gradient import split_evol_ops, _split_evol_tapes, stoch_pulse_grad
 
 
-def test_stoch_pulse_grad_raises_without_jax_installed():
-    """Test that an error is raised if a convenience function is called without jax installed"""
-    try:
-        # Need to import a specific subpackage of JAX, because `import jax` would import the
-        # module `tests/returntypes/jax`.
-        from jax import random  # pylint: disable=unused-import
-
-        pytest.skip()
-    except ImportError:
-        tape = qml.tape.QuantumScript([], [])
-        with pytest.raises(ImportError, match="Module jax is required"):
-            stoch_pulse_grad(tape)
-
 
 @pytest.mark.jax
 class TestSplitEvolOps:
